+++ packages/shared/src/index.ts
@@ -1,2 +1,3 @@
 export * from './fs'
 export * from './date'
+export * from './vitepress'