+++ packages/theme/src/components/BlogBackToTop.vue
@@ -1,9 +1,8 @@
 <script lang="ts" setup>
 import { useElementSize, useScroll } from '@vueuse/core'
-import { ElIcon } from 'element-plus'
 import { computed, ref } from 'vue'
-import { vOuterHtml } from '../directives'
-import { useBackToTopConfig } from '../composables/config/blog'
+import { useBackToTopConfig, useOpenBackToTop } from '../composables/config/blog'
+import Icon from './Icon.vue'
 
 function handleBackRoTop() {
   window.scrollTo({ top: 0, behavior: 'smooth' })
@@ -15,29 +14,28 @@ const { width } = useElementSize(el)
 const docWidth = computed(() => `${width.value}px`)
 
 const backToTopConfig = useBackToTopConfig()
-const open = computed(() => !!(backToTopConfig ?? true))
+const open = useOpenBackToTop()
 
 const { y } = useScroll(window)
 const defaultTriggerHeight = 450
-const triggerTop = computed(() => (typeof backToTopConfig === 'boolean' ? undefined : backToTopConfig?.top) ?? defaultTriggerHeight)
+const triggerTop = computed(() => backToTopConfig.value?.top ?? defaultTriggerHeight)
 
 const show = computed(() => width && y.value > triggerTop.value)
 
-const iconSVGStr = computed(() => typeof backToTopConfig === 'boolean' ? '' : backToTopConfig?.icon)
+const iconSVGStr = computed(() => backToTopConfig?.value?.icon)
 </script>
 
 <template>
   <div v-if="open" v-show="show" class="back-to-top">
     <span class="icon-wrapper" @click="handleBackRoTop">
-      <ElIcon :size="20">
-        <i v-if="iconSVGStr" v-outer-html="iconSVGStr" />
-        <svg v-else width="512" height="512" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+      <Icon :size="20" :icon="iconSVGStr">
+        <svg width="512" height="512" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
           <path
             fill="currentColor"
             d="m20 22l-3.86-1.55c.7-1.53 1.2-3.11 1.51-4.72zM7.86 20.45L4 22l2.35-6.27c.31 1.61.81 3.19 1.51 4.72M12 2s5 2 5 10c0 3.1-.75 5.75-1.67 7.83A2 2 0 0 1 13.5 21h-3a2 2 0 0 1-1.83-1.17C7.76 17.75 7 15.1 7 12c0-8 5-10 5-10m0 10c1.1 0 2-.9 2-2s-.9-2-2-2s-2 .9-2 2s.9 2 2 2"
           />
         </svg>
-      </ElIcon>
+      </Icon>
     </span>
   </div>
 </template>