+++ packages/theme/src/components/BlogSidebar.vue
@@ -1,35 +1,29 @@
 <script lang="ts" setup>
 import { computed } from 'vue'
-import { useBlogConfig } from '../composables/config/blog'
+import { useRecommendConfig, useShowRecommend } from '../composables/config/blog'
 import BlogRecommendArticle from './BlogRecommendArticle.vue'
 
-const { recommend: _recommend } = useBlogConfig()
-
+const recommend = useRecommendConfig()
+const show = useShowRecommend()
 const sidebarStyle = computed(() =>
-  _recommend && _recommend?.style ? _recommend.style : 'card'
-)
-const marginTop = computed(() =>
-  sidebarStyle.value === 'card' ? '40px' : '0px'
-)
-const marginTopMini = computed(() =>
-  sidebarStyle.value === 'card' ? '60px' : '0px'
+  recommend.value?.style || 'sidebar'
 )
 </script>
 
 <template>
-  <div v-if="_recommend !== false" class="sidebar" data-pagefind-ignore="all">
+  <div v-if="show" class="sidebar" :class="{ card: sidebarStyle === 'card' }" data-pagefind-ignore="all">
     <BlogRecommendArticle />
   </div>
 </template>
 
 <style lang="scss" scoped>
-.sidebar {
-  margin-top: v-bind(marginTop);
+.sidebar.card {
+  margin-top: 40px;
 }
 
 @media screen and (min-width: 960px) and (max-width: 1120px) {
-  .sidebar {
-    margin-top: v-bind(marginTopMini);
+  .sidebar.card {
+    margin-top: 60px;
   }
 }
 </style>