+++ packages/vitepress-plugin-rss/CHANGELOG.md
@@ -1,5 +1,13 @@
 # vitepress-plugin-rss
 
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @sugarat/theme-shared@0.0.3
+- feat: support rewrites and dynamic routes
+
 ## 0.2.10
 
 ### Patch Changes